requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
dearpygui==1.9.1
Django==4.2.2
django-phonenumber-field==7.1.0
djangorestframework==3.14.0
numpy==1.23.2
opencv-contrib-python==4.6.0.66
packaging==21.3
phonenumbers==8.13.14
Pillow==9.2.0
pyFirmata==1.1.0
pyparsing==3.0.9
pyserial==3.5
pytesseract==0.3.10
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
